8239334: Tab Size does not work correctly in JTextArea with setLineWrap on

+++ src/java.desktop/share/classes/javax/swing/text/WrappedPlainView.java
@@ -466,7 +466,7 @@ public class WrappedPlainView extends BoxView implements TabExpander {
     public float nextTabStop(float x, int tabOffset) {
         if (tabSize == 0)
             return x;
-        float ntabs = (x - tabBase) / tabSize;
+        int ntabs = (int) ((x - tabBase) / tabSize);
         return tabBase + ((ntabs + 1) * tabSize);
     }
